// src/app/models/employeesalaryrequest.ts
export class EmployeeSalaryRequest {
    id: number;
    employeeId: number;
    salaryAmount: number;
    retentionBonusPerMonth: number;
    retentionBonusPayment: number;
    retentionPaymentSalaryMonth: string;
    nextAppraisalSalaryMonth: string;
    salaryStart: any;
    salaryEnd: any;
    isActive: string;
    range: any;
    leaveStartDate: any;
    leaveEndDate: any;
    modifiedBy: number;
    createdBy: number;
}